requirements.txt
SQLAlchemy~=2.0.25
fastapi~=0.110.0
uvicorn~=0.27.1
python-telegram-bot[rate-limiter]~=20.7
python-telegram-bot[webhooks]~=20.7
python-telegram-bot[job-queue]~=20.7
redis~=5.0.1
mysql-connector-python~=8.3.0
